rename all index files to files that can easily be found without typing too much. Index files are harder to find by open file and can cause confusion in editors.
