+++ build-tools/config/webpack/webpack.base.conf.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const config = require('../../config');
+const config = require('../../config/config');
 const webpackHelpers = require('./webpackHelpers');
 
 const projectRoot = path.resolve(__dirname, '../../../');
@@ -8,7 +8,7 @@ const isDevelopment = process.env.NODE_ENV === 'development';
 module.exports = {
   entry: {
     app: [
-      './src/polyfill/index.js',
+      './src/polyfill/polyfill.js',
       './src/bootstrap.js'
     ],
   },
