+++ build-tools/config/webpack/webpack.dev.conf.js
@@ -1,4 +1,4 @@
-const config = require('../../config');
+const config = require('../../config/config');
 const webpack = require('webpack');
 const merge = require('webpack-merge');
 const baseWebpackConfig = require('./webpack.base.conf');
