+++ build-tools/script/preview-server.js
@@ -1,4 +1,4 @@
-const config = require('../config');
+const config = require('../config/config');
 if (!process.env.NODE_ENV) process.env.NODE_ENV = JSON.parse(config.dev.env.NODE_ENV);
 const path = require('path');
 const express = require('express');
